changes to respond_get
