--- core/src/network/reducers/respond_get.rs
@@ -1,92 +1,0 @@
-use crate::{
-    action::{ActionWrapper, RespondQueryPayload},
-    network::{
-        actions::ActionResponse,
-        query::{GetLinksNetworkResult, NetworkQueryResult},
-        reducers::send,
-        state::NetworkState,
-    },
-    state::State,
-};
-use holochain_core_types::{entry::EntryWithMetaAndHeader, error::HolochainError};
-use holochain_json_api::json::JsonString;
-use lib3h_protocol::{
-    data_types::{QueryEntryData, QueryEntryResultData},
-    protocol_client::Lib3hClientProtocol,
-};
-
-/// Send back to network a HandleQueryEntryResult, no matter what.
-/// Will return an empty content field if it actually doesn't have the data.
-fn reduce_respond_get_links_inner(
-    network_state: &mut NetworkState,
-    query_data: &QueryEntryData,
-    links: &GetLinksNetworkResult,
-    link_type: String,
-    tag: String,
-) -> Result<(), HolochainError> {
-    network_state.initialized()?;
-    let query_result_json: JsonString =
-        NetworkQueryResult::Links(links.clone(), link_type, tag).into();
-    send(
-        network_state,
-        Lib3hClientProtocol::HandleQueryEntryResult(QueryEntryResultData {
-            request_id: query_data.request_id.clone(),
-            requester_agent_id: query_data.requester_agent_id.clone(),
-            space_address: network_state.dna_address.clone().unwrap(),
-            responder_agent_id: network_state.agent_id.clone().unwrap().into(),
-            entry_address: query_data.entry_address.clone().into(),
-            query_result: query_result_json.to_string().into_bytes(),
-        }),
-    )
-}
-
-/// Send back to network a HandleQueryEntryResult, no matter what.
-/// Will return an empty content field if it actually doesn't have the data.
-fn reduce_respond_get_inner(
-    network_state: &mut NetworkState,
-    query_data: &QueryEntryData,
-    maybe_entry: &Option<EntryWithMetaAndHeader>,
-) -> Result<(), HolochainError> {
-    network_state.initialized()?;
-    let query_result_json: JsonString = NetworkQueryResult::Entry(maybe_entry.clone()).into();
-    send(
-        network_state,
-        Lib3hClientProtocol::HandleQueryEntryResult(QueryEntryResultData {
-            request_id: query_data.request_id.clone(),
-            requester_agent_id: query_data.requester_agent_id.clone(),
-            space_address: network_state.dna_address.clone().unwrap(),
-            responder_agent_id: network_state.agent_id.clone().unwrap().into(),
-            entry_address: query_data.entry_address.clone().into(),
-            query_result: query_result_json.to_string().into_bytes(),
-        }),
-    )
-}
-
-pub fn reduce_respond_get(
-    network_state: &mut NetworkState,
-    _root_state: &State,
-    action_wrapper: &ActionWrapper,
-) {
-    let action = action_wrapper.action();
-    let (query_data, payload) = unwrap_to!(action=>crate::action::Action::RespondQuery);
-    let result = match payload {
-        RespondQueryPayload::Entry(maybe_entry) => {
-            reduce_respond_get_inner(network_state, query_data, maybe_entry)
-        }
-        RespondQueryPayload::Links((links, link_type, tag)) => reduce_respond_get_links_inner(
-            network_state,
-            query_data,
-            links,
-            link_type.clone(),
-            tag.clone(),
-        ),
-    };
-
-    let result = result
-        .map(|_| Ok(()))
-        .unwrap_or_else(|e| Err(HolochainError::ErrorGeneric(e.to_string())));
-
-    network_state
-        .actions
-        .insert(action_wrapper.clone(), ActionResponse::Respond(result));
-}
